requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Werkzeug==2.3.7
matplotlib==3.8.0
numpy==1.26.0
reportlab==4.0.9
python-dotenv==1.0.0
